services/ml-prediction/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
xgboost==2.0.2
joblib==1.3.2
mysql-connector-python==8.2.0
redis==5.0.1
python-dotenv==1.0.0
